Initial commit - JAX-RS tutorial part1 without server
This is Eclipse project and it can be readily imported into Eclipse.

Initial commit as part of JAX-RS tutorial by Christophe Coenraets.
This mini-project has started to simplify the source code of Christophe.

Tutorial link:
http://coenraets.org/blog/2011/12/backbone-js-wine-cellar-tutorial-part-1-getting-started/

Tutorial git code link:
https://github.com/ccoenraets/backbone-cellar/tree/master/tutorial/part1

The JavaScript code works till it reaches this.wineList.fetch() in main.js, then
it fails (as it should be), because there is no server to answer
the GET request:

GET http://localhost:8080/WineCellar/api/wines